Add logic for modal state update
